refactor: 슬롯 검증 로직을 ReservationSlot으로 위임
